Fix Serialization Regression from Upgrade to Jackson 2
The `StdObjectMapperFactory` is now calling `setSerializationInclusion`
directly on the `ObjectMapper` to ignore null fields in the default
configuration.

In Jackson 1, it was OK to modify the `SerializationConfig` of an
`ObjectMapper`. In Jackson 2, the `SerializationConfig` is immutable
and calling the `with...` methods produces a new copy of the object that
can no longer be assigned to an `ObjectMapper`.

https://fasterxml.github.io/jackson-databind/javadoc/2.3.0/com/fasterxml/jackson/databind/SerializationConfig.html